move script to prisma update
